dependency builds use Cargo
With this change, individual submodules can specify their dependencies with
an additional file called "deps.mk" in the subdir. When building, only
the dependencies that are necessary are built, using cargo, and then linked.

This greatly simplifies adding new dependencies: add the package in
deps/Cargo.toml, and add the appropriate line in "deps.mk" in the
src/utilname/ directory, and the dependency will be built automatically
as needed.

This also removes the need to use git submodules.

+++ src/fmt/linebreak.rs
@@ -10,7 +10,6 @@
 use FmtOptions;
 use parasplit::{Paragraph, ParaWords, WordInfo};
 use std::io::{Write, BufWriter, Stdout};
-use std::num::{Float, Int, SignedInt};
 use std::i64;
 use std::cmp;
 use std::mem;
@@ -373,7 +372,7 @@ fn compute_demerits(delta_len: isize, stretch: isize, wlen: isize, prev_rat: f32
     // we penalize lines that have very different ratios from previous lines
     let bad_delta_r = (DR_MULT * (((ratio - prev_rat) / 2.0).powf(3f32)).abs()) as i64;
 
-    let demerits = Int::pow(1 + bad_linelen + bad_wordlen + bad_delta_r, 2);
+    let demerits = i64::pow(1 + bad_linelen + bad_wordlen + bad_delta_r, 2);
 
     (demerits, ratio)
 }
